app/requirements.txt
boto==2.49.0
boto3==1.24.42
botocore==1.27.42
jmespath==1.0.1
mangum==0.15.0
pydantic==1.9.1
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==5.4
s3transfer==0.6.0
six==1.16.0
tornado==6.2
tornado-swirl==0.1.20
typing_extensions==4.3.0
urllib3==1.26.11
